add change tests
